buffio queue changes

+++ buffio.hpp
@@ -23,6 +23,7 @@ enum BUFFIO_SOCK_TYPE{
  BUFFIO_SOCK_TCP = SOCK_STREAM,
  BUFFIO_SOCK_UDP = SOCK_DGRAM,
  BUFFIO_SOCK_RAW = SOCK_RAW,
+ BUFFIO_SOCK_ASYNC = SOCK_NONBLOCK, 
 };
 
 //user can give number between MAX and MINIMUM;
@@ -39,28 +40,44 @@ enum BUFFIO_ROUTINE_STATUS{
   BUFFIO_ROUTINE_YIELD,
 };
 
-struct buffioinfo{
- const char *address;
- int portnumber;
- int listenbacklog;
- size_t capacity;
- size_t reserve;
- BUFFIO_SOCK_TYPE socktype;
- BUFFIO_FAMILY_TYPE sockfamily;
+enum BUFFIO_QUEUE_OVERFLOW_POLICY{
+ OVERFLOWPOLICY_NONE,
+ OVERFLOWPOLICY_ALLOCATE_NEW,
+ OVERFLOWPOLICY_DISCARD_ROUTINE,
+ OVERFLOWPOLICY_EXEC_AFTER_FREE,
 };
 
+enum BUFFIO_QUEUE_CHUNKALLOCATION_POLICY{
+ CHUNKALLOCATION_POLICY_NONE,
+ CHUNKALLOCATION_POLICY_NEW_CHUNK,
+ CHUNKALLOCATION_POLICY_SIZE_DOUBLE,
+};
 
-struct buffiopromise;
-struct buffioqueue;
-struct buffiobuffer;
-struct clientinfo;
-struct acceptreturn;
-using buffiopromise = struct buffiopromise;
-using buffioinfo = struct buffioinfo;
+enum BUFFIO_QUEUE_THREAD_POLICY{
+ THREAD_POLICY_PER_CHUNK,
+ THREAD_POLICY_NONE,
+};
 
-#define iowait co_await
-#define ioyeild co_yield
-#define ioreturn co_return
+enum BUFFIO_QUEUE_ROUTINE_ERROR_POLICY{
+ ROUTINE_ERROR_POLICY_NONE,
+ ROUTINE_ERROR_POLICY_SHUTDOWN,
+ ROUTINE_ERROR_POLICY_CONTINUE,
+ ROUTINE_ERROR_POLICY_CALLBACK,
+};
+
+enum BUFFIO_EVENTLOOP_TYPE{
+  EVENTLOOP_SYNC,
+  EVENTLOOP_ASYNC,
+};
+
+enum BUFFIO_ACCEPT_STATUS{
+  BUFFIO_ACCEPT_STATUS_ERROR = -1,
+  BUFFIO_ACCEPT_STATUS_SUCCESS,
+  BUFFIO_ACCEPT_STATUS_NA,
+  BUFFFIO_ACCEPT_STATUS_NO_HANDLER,
+};
+
+#if defined(BUFFIO_IMPLEMENTATION)
 
 namespace buffio{
  class buffsocket;
@@ -69,18 +86,31 @@ namespace buffio{
  class queue;
 };
 
-#if defined(BUFFIO_IMPLEMENTATION)
+struct buffioinfo{
+ const char *address;
+ int portnumber;
+ int listenbacklog;
+ int socktype;
+ BUFFIO_FAMILY_TYPE sockfamily;
+};
 
-// for public use;
+struct buffioqueuepolicy{
+ enum BUFFIO_QUEUE_OVERFLOW_POLICY overflowpolicy;
+ enum BUFFIO_QUEUE_THREAD_POLICY threadpolicy;
+ enum BUFFIO_QUEUE_CHUNKALLOCATION_POLICY chunkallocationpolicy;
+ enum BUFFIO_QUEUE_ROUTINE_ERROR_POLICY routineerrorpolicy;
+ int queuecapacity;
+};
+
 struct buffiopromise;
+using buffiopromise = struct buffiopromise;
+using buffioinfo = struct buffioinfo;
+using buffioqueuepolicy = struct buffioqueuepolicy;
 using buffiohandleroutine = std::coroutine_handle<buffiopromise>;
 
-struct buffioqueue{
- buffiohandleroutine handle;
- buffiohandleroutine complitioncallback;
- int priority;
- int priorityexpire; // used to control how much subsiquent call has to be made to this thing
-};
+#define iowait co_await
+#define ioyeild co_yield
+#define ioreturn co_return
 
 
 struct buffiobuffer{
@@ -106,6 +136,7 @@ struct acceptreturn{
  int errorcode;
  buffioroutine handle;
 };
+
 
 struct buffioawaiter{
    bool await_ready() const noexcept{ return false;}
@@ -149,19 +180,12 @@ class buffio::buffsocket{
 
  public:
  
-  buffsocket(buffioinfo &ioinfo) : linfo(ioinfo), clienthandler(nullptr){
-     if(ioinfo.capacity < 1){
-       BUFFIO_ERROR(" SOCKET: Field maxclient in buffioinfo structure cannot be less than 1");
-     return;
-     };
+  buffsocket(buffioinfo &ioinfo) : linfo(ioinfo), clienthandler(nullptr),
+                                   socketfd(-1) , sockfdblocking(false){
      switch(ioinfo.sockfamily){
-      case BUFFIO_FAMILY_LOCAL:
-        if(createlocalsocket(ioinfo,&socketfd) < 0)
-           return;
+      case BUFFIO_FAMILY_LOCAL: if(createlocalsocket(ioinfo,&socketfd) < 0) return;
       break;
-      case BUFFIO_FAMILY_IPV4:
-        if(createipv4socket(ioinfo,&socketfd) < 0)
-            return;
+      case BUFFIO_FAMILY_IPV4:  if(createipv4socket(ioinfo,&socketfd) < 0) return;
       break;
       case BUFFIO_FAMILY_IPV6:
       break;
@@ -173,11 +197,12 @@ class buffio::buffsocket{
       break;
       case BUFFIO_FAMILY_BLUETOOTH:
       break;
-      default:
-       BUFFIO_ERROR(" UNKNOWN TYPE SOCKET CREATION FAILED");
-      return;
+      default: BUFFIO_ERROR(" UNKNOWN TYPE SOCKET CREATION FAILED "); return;
       break;
-    }; 
+    };
+    
+    if(!(ioinfo.socktype & BUFFIO_SOCK_ASYNC)) sockfdblocking = true;
+             
      return;
   };
   
@@ -222,7 +247,8 @@ class buffio::buffsocket{
       if(close(socketfd) < 0){
         BUFFIO_ERROR(" SOCKET CLOSE FAILURE");
      };
-     unlink(linfo.address);
+     if(linfo.sockfamily  == BUFFIO_FAMILY_LOCAL) unlink(linfo.address);
+
      BUFFIO_LOG(" SOCKET CLOSED SUCCESFULLY");
     }
  };
@@ -248,32 +274,36 @@ acceptreturn acceptsock(){
            sockinfolen = sizeof(struct sockaddr_in); 
            struct sockaddr_in address_in;
            afd = accept(socketfd,(struct sockaddr*)&address_in,&sockinfolen);
-           portnumber = ntohs(address_in.sin_port);
-           aacceptedaddress  = inet_ntoa((struct in_addr)address_in.sin_addr); 
+           if(afd > 0){
+             portnumber = ntohs(address_in.sin_port);
+             aacceptedaddress  = inet_ntoa((struct in_addr)address_in.sin_addr); 
+           }
          }
         break;
         case BUFFIO_FAMILY_IPV6:  break;
      }; 
 
-
-      if(afd  < 0){
-         BUFFIO_ERROR(" FAILED TO ACCEPT A CONNECTION \n reason : ",strerror(errno));
-         return {.errorcode = -1 , .handle = NULL};
-      }
+    if(afd < 0){
+       if(errno == EAGAIN || errno == EWOULDBLOCK)
+             return {.errorcode = BUFFIO_ACCEPT_STATUS_NA , .handle = NULL};
       
-      if(!clienthandler) return {.errorcode = 1,.handle = NULL};
-
+       BUFFIO_ERROR(" Failed to accept connection, \n reason: ",strerror(errno),errno);
+             return {.errorcode = BUFFIO_ACCEPT_STATUS_ERROR , .handle = NULL};
+    }
+  
       cinfo = {.address = aacceptedaddress , .clientfd = afd,.portnumber = portnumber};
-
-      return {.errorcode = 0 , .handle = clienthandler(cinfo)};
+      if(!clienthandler) return {.errorcode = BUFFFIO_ACCEPT_STATUS_NO_HANDLER,.handle = NULL};
+      return {.errorcode = BUFFIO_ACCEPT_STATUS_SUCCESS , .handle = clienthandler(cinfo)};
 };
  
- int socketfd = -1;
+ int socketfd;
+ bool sockfdblocking;
+
  buffioinfo linfo;
  buffioroutine (*clienthandler)(clientinfo cinfo);
+ clientinfo cinfo;
 
  private:
-      clientinfo cinfo;
       socklen_t sockinfolen = 0;
       char *aacceptedaddress = nullptr;
       int afd = -1;
@@ -281,95 +311,211 @@ acceptreturn acceptsock(){
 };
 
 
+struct buffiotaskchunk{
+       int priority{0};
+       int busychunk{0};
+       buffioroutine handle;
+       buffioroutine complitioncallback;
+       buffiotaskchunk *next{nullptr};
+       buffiotaskchunk *prev{nullptr};
+};
+
+class buffioqueuechunk{
+
+public:
+  struct buffiotaskchunk *allocatedchunk;
+  struct buffiotaskchunk *freetaskchunk;
+  struct buffiotaskchunk *execqueue;
+  struct buffiotaskchunk *execqueuetail;
+  size_t currentat;
+  size_t chunkcapacity;
+  size_t chunkfilled;
+  struct buffioqueuechunk *next;
+  struct buffioqueuechunk *prev;
+
+  bool isfull(){ return chunkfilled < chunkcapacity ? false : true;}
+
+  buffioqueuechunk(size_t capacity) : 
+      chunkcapacity(capacity), chunkfilled(0),
+      allocatedchunk(nullptr), freetaskchunk(nullptr),
+      execqueue(nullptr), execqueuetail(nullptr),
+      currentat(0),next(nullptr),prev(nullptr)
+  {
+     allocatedchunk = new buffiotaskchunk[capacity];
+     freetaskchunk = allocatedchunk;
+     struct buffiotaskchunk *head = allocatedchunk;
+       
+    for(int i = 1; i < capacity; i++){
+     head->next = allocatedchunk + i;
+     allocatedchunk[i].prev = head;
+     head = head->next;
+    }
+    BUFFIO_TRACE(" Allocated task chunk for queue chunk, ptraddress : ",allocatedchunk);
+    BUFFIO_LOG(" Task queue chunk is ready.");
+    BUFFIO_INFO(" Task queue chunk OK. ");
+  };
+
+  void clean(){ delete this->allocatedchunk; }
+  void pushtask(buffioroutine routine){
+   BUFFIO_TRACE(" Pushing routine to task queue");
+   freetaskchunk->handle = routine;
+   execqueuetail->next = freetaskchunk;
+   execqueuetail = freetaskchunk;
+   if(freetaskchunk->next)
+       freetaskchunk = freetaskchunk->next;
+
+    chunkfilled = (chunkfilled < chunkcapacity) ? chunkfilled  + 1 : chunkcapacity;
+   BUFFIO_TRACE(" Pushed routine to task queue");
+
+  };
+  void execute(){
+   if(execqueue->handle)
+     return;
+  };
+};
+
+//using buffiotaskchunk = struct buffiotaskchunk;
+using buffioqueuechunk = struct buffioqueuechunk;
+
 class buffio::queue{
   public:
-    queue(size_t queuesize,size_t reservesize) :  capacity(queuesize), 
-             occupiedcapacity(0), currentidx(0),
-             execqueue(nullptr), emptyplaces(nullptr),
-             reservesize(reservesize)
+   
+    queue(buffioqueuepolicy &queuepolicy) :  capacity(0), 
+             occupiedcapacity(0), chunksnumber(0),
+             chunkhead(nullptr),chunknext(nullptr),
+             queuepolicy(queuepolicy),queueerror(0),
+             chunkcurrent(nullptr)
    { 
-       if(queuesize < 1){
-         BUFFIO_ERROR(" QUEUE: field queuesize to queue is less than 1," 
-                          " it must have to be 1 or greater than 1");
-        return;
-       }
-       if(reservesize < 10){
-         BUFFIO_ERROR(" QUEUE: field reservesize to queue is less than 10,"
-                          " using no reserve policy");
-        reservesize = 0;
-       } 
-       execqueue = new buffioqueue[queuesize + reservesize];
-       emptyplaces = new buffioqueue*[queuesize + reservesize];
-             
-       if(!execqueue || ! emptyplaces){
-           BUFFIO_ERROR(" QUEUE CREATION FAILED ");
-           BUFFIO_DEBUG(" Allocation of one of the queue failed ");
-           BUFFIO_TRACE(" EXEC QUEUE : ",execqueue," QUEUE emptytracker",emptyplaces);
-        return; 
+       if(queuepolicy.queuecapacity < 1){
+         BUFFIO_ERROR(" QUEUE: field queuecapacity to queue is less than 1," 
+                      " it must have to be 1 or greater than 1");
+         queueerror = -1;
+         return;
        }
 
-       reservequeue = reservesize < 10 ? nullptr : (execqueue + (queuesize - 1));
-       current = execqueue;
-       Next = execqueue + 1;
+       chunkhead = new buffioqueuechunk(queuepolicy.queuecapacity);
+//       chunkhead->initchunk(queuepolicy.queuecapacity);
+       chunkempty = chunkhead;
+       chunkcurrent = chunkhead;
+       chunkhead->next = nullptr;
+       chunknext = chunkhead;
+       chunksnumber += 1;
+       BUFFIO_INFO(" Queue full, cannot push client handler.");
+
    };
 
  ~queue(){
-    if(execqueue) delete execqueue;
-    if(emptyplaces) delete emptyplaces;
-  };
+   struct buffioqueuechunk *next;
+   for(auto *itr = chunkhead ; itr != nullptr ; ){
+       itr->clean();
+       next = itr->next;
+       delete itr;
+       itr = next;
+    }
+        
+};
 
  int push(buffioroutine routine){
-    if(!(occupiedcapacity < capacity)){
+    if(!chunkempty->isfull()){
+         BUFFIO_INFO(" Pushed routine to queue"); 
+         chunkempty->pushtask(routine);
 
-           BUFFIO_WARN(" QUEUE: cannot push instance to queue is full. execution will continue"
-                           " when queue is free");
-     }
-     routine.resume(); 
+      return 0;
+    };
+
+    switch(queuepolicy.overflowpolicy){
+        case OVERFLOWPOLICY_NONE:{
+         BUFFIO_INFO(" Queue full, cannot push client handler.");
+         BUFFIO_LOG(" Current Queue Policy : OVERFLOWPOLICY_NONE.");
+         BUFFIO_TRACE(" Destroying Routine Handle");
+         routine.destroy();
+        } break;
+
+        case OVERFLOWPOLICY_ALLOCATE_NEW:{ 
+         BUFFIO_INFO("All queuechunk busy, Allocating new queue chunk");
+         buffioqueuechunk *tmpchunk = new buffioqueuechunk(queuepolicy.queuecapacity);
+         chunknext->next = tmpchunk;
+//         tmpchunk->initchunk(queuepolicy.queuecapacity);
+         chunknext = tmpchunk; 
+         BUFFIO_TRACE(" Current Queue Policy : OVERFLOWPOLICY_DISCARD_ROUTINE. chunk : ",tmpchunk);
+
+        }break; 
+        case OVERFLOWPOLICY_DISCARD_ROUTINE:{ 
+                BUFFIO_INFO(" Cannot push client to queue, queuepolicy restrict that.");
+                BUFFIO_LOG(" Current Queue Policy : OVERFLOWPOLICY_DISCARD_ROUTINE.");
+                BUFFIO_TRACE(" Destroying Routine Handle");
+                routine.destroy();                 
+         } break;
+        case OVERFLOWPOLICY_EXEC_AFTER_FREE:{
+
+        } break;
+
+    };
     return 0;
  };
 
   void yield(){
-     currentidx += 1;
-      size_t nidx = (currentidx < capacity) * currentidx;
-      if(!execqueue[currentidx].handle){
-         BUFFIO_TRACE(" NO ROUTINE AVALILABLE TO EXECUTE ");
-         return;
-      }
-    execqueue[currentidx].handle.resume();
+      chunkcurrent->execute();
     return;
   };
-   
+
+  int queueerror;
+  
  private:
 
- buffioqueue *execqueue;
- buffioqueue *reservequeue;
- buffioqueue **emptyplaces;
- buffioqueue  *current;
- buffioqueue  *Next;
+ buffioqueuechunk *chunkhead ,*chunknext;
+ buffioqueuechunk *chunkcurrent, *chunkempty;
+ buffioqueuechunk *chunkexecfree, *chunkexecfreetail;
 
- size_t emptyptridx;
+ buffioqueuepolicy queuepolicy;
+
+ size_t chunksnumber;
  size_t capacity;
  size_t occupiedcapacity;
- size_t reservesize;
- size_t currentidx;
 };
 
 class buffio::instance{
 
 public:
   
+  static int eventloop(void *data){
+      buffio::instance *instance = (buffio::instance *)data;
+      buffio::queue queue(instance->instancequeuepolicy);   
+      if(queue.queueerror < 0)
+         return 0;
 
-  void fireeventloop(buffsocket &sock){
-     buffio::queue queue(sock.linfo.capacity , sock.linfo.reserve);
+     acceptreturn accepted = instance->instancesock->acceptsock();
+     switch(accepted.errorcode){
+      case  BUFFIO_ACCEPT_STATUS_ERROR: BUFFIO_LOG(" Error while accepting a connection");
+      case  BUFFFIO_ACCEPT_STATUS_NO_HANDLER: queue.yield(); break;
+      case  BUFFIO_ACCEPT_STATUS_SUCCESS: queue.push(accepted.handle); break;
+      case  BUFFIO_ACCEPT_STATUS_NA: queue.yield(); break;
+      default: queue.yield();
+     };
+  return 0;
+  }
+  
+  void fireeventloop(buffsocket &sock,buffioqueuepolicy &queuepolicy
+                     ,enum BUFFIO_EVENTLOOP_TYPE eventlooptype){
+
+     instancequeuepolicy = queuepolicy;
+     instancesock = &sock;
      sock.listensock();
-     acceptreturn accepted = sock.acceptsock();
-     queue.push(accepted.handle);
-     queue.yield(); 
-    return;
+     switch(eventlooptype){
+     case EVENTLOOP_SYNC: eventloop(this); break; 
+     case EVENTLOOP_ASYNC: break;
+     }
+
+     return;
   };
+
   ~instance(){
    //clean up-code;
   };
+
+private:
+  buffsocket *instancesock;
+  buffioqueuepolicy instancequeuepolicy;
 };
 
 #endif
